+++ src/read.rs
@@ -10,12 +10,12 @@ pub(crate) fn parse_data<E: fmt::Debug>(buf: &[u8; PAYLOAD_LEN]) -> Result<Readi
     let expected_sum: u16 = ((buf[PAYLOAD_LEN-2] as u16) << 8) | (buf[PAYLOAD_LEN-1] as u16);
     if expected_sum == sum {
         Ok(Reading {
-            pm1: as_u16(buf[4], buf[5]) as f32,
-            pm2_5: as_u16(buf[6], buf[7]) as f32,
-            pm10: as_u16(buf[8], buf[9]) as f32,
-            env_pm1: as_u16(buf[10], buf[11]) as f32,
-            env_pm2_5: as_u16(buf[12], buf[13]) as f32,
-            env_pm10: as_u16(buf[14], buf[15]) as f32,
+            pm1: as_u16(buf[4], buf[5]),
+            pm2_5: as_u16(buf[6], buf[7]),
+            pm10: as_u16(buf[8], buf[9]),
+            env_pm1: as_u16(buf[10], buf[11]),
+            env_pm2_5: as_u16(buf[12], buf[13]),
+            env_pm10: as_u16(buf[14], buf[15]),
             particles_0_3: as_u16(buf[16], buf[17]),
             particles_0_5: as_u16(buf[18], buf[19]),
             particles_1: as_u16(buf[20], buf[21]),
